refactor: systematic dead-code elimination sweep
- Remove dead TOTP code (internal/auth/totp.go, totp_test.go) and associated coverage tests
- Delete entire internal/ingress/middleware package (compress, ratelimit, ratelimit_bolt) — unused in production
- Remove dead CircuitBreakerRegistry from internal/core/circuitbreaker.go and tests
- Remove dead Build/CloneGitRepo/DetectBuildPack from internal/topology/deployer.go and tests
- Fix errcheck warnings in cmd/deploymonster/main.go and tests/soak/main.go
- Fix unnecessary conversions in resources.go and metrics_export.go
- Fix empty branch warning in backup/s3.go

+++ internal/api/handlers/metrics_export.go
@@ -58,7 +58,7 @@ func (h *MetricsExportHandler) Export(w http.ResponseWriter, r *http.Request) {
 				stats, err := h.runtime.Stats(r.Context(), containers[0].ID)
 				if err == nil {
 					currentCPU = stats.CPUPercent
-					currentMemMB = int64(stats.MemoryUsage / (1024 * 1024))
+					currentMemMB = stats.MemoryUsage / (1024 * 1024)
 				}
 			}
 		}
